Reorganizing test files
